# Optimize plan creation when wisdom already exists